update linux2 build scripts, using ubuntu 12.04
ubuntu 10.04 can't run in docker, crashed when call bash

+++ kdev.py
@@ -438,9 +438,11 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-diff .config .config-bak
-if [ $? -eq 0 ];then
-    cp -a .config-bak .config
+if [ -f .config-bak ] ; then
+    diff .config .config-bak
+    if [ $? -eq 0 ];then
+        cp -a .config-bak .config
+    fi
 fi
 
 ls -alh ${WORKDIR}/.config
